case where the socket has been closed before we read headers
catch the NoMoreData errors raised when parsing the headers. By defaut
send doesn return an error if the remote socket is closed. instead the
system zill just ignore it. But zhen ze try to read the response then
the NoMoreData error was raised. This change catch the error and make
sure ze return in the retry loop if we can.